SEGURIDAD SOLUCIONADO

+++ EDIA/src/main/java/pe/upc/edu/edia/servicesinterfaces/IUsuarioService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IUsuarioService {
     public List<Usuario> list();
     public void insert(Usuario u);
-    public Usuario listId(Long idUsuario);
+    public Usuario listId(int idUsuario);
     public void update(Usuario u);
-    public void delete(Long idUsuario);
+    public void delete(int idUsuario);
 }
